addressed linting errors fixed database typo and brokedown large file of SplatViewer
