+++ app/db/schema.ts
@@ -4,7 +4,7 @@ import {pgTable, serial, text, timestamp} from  'drizzle-orm/pg-core';
 export const splats=pgTable('splats',{
     id: serial('id').primaryKey(),
     name: text('name'),
-    spalt:text('splat'),
+    splat:text('splat'),
     video:text('video'),
     createdAt:timestamp('createdAt').defaultNow(),
     updatedAt:timestamp('updatedAt')
